+++ src/traffic_generator_templates.cc
@@ -8,7 +8,7 @@ void TrafficGenerator::tick(const uint64_t tickno, NextHop & next) {
   for (uint16_t i = 0; i < num_ports_; i++) {
     uint16_t output_port = i;
     while (output_port == i) {
-      output_port = distribution_(prng_);
+      output_port = static_cast<uint16_t>(distribution_(prng_));
     }
     Cell cell(i, output_port);
     next.accept(tickno, cell);
